refactor(model): Rename samples history module
